fix(cli): remove parsed-but-ignored flags (AUDIT §5(b)5)
Remove eight clap flags that parsed but were silently discarded — an
operator trap where commands pretended to honor options they ignored:

  shards --node, record --slots, record --raw, replication --history,
  redo --tail, rebalance --execute, drain --cancel, log-level --target

None had server-side backing: /debug/records returns metadata only,
/admin/replication has no history, /debug/redo has no entry listing,
rebalance executes by default, drain has no cancel endpoint, and
PUT /debug/log-level only accepts a global level. Implementing any
would require new endpoints, so removal is the honest fix.

Update specs/BSV_UTXO_STORE_SPEC.md §14.2 and phases/13 checklist to
match the real CLI surface. Add cli_removed_noop_flags_are_rejected
proving clap now rejects each removed flag with a usage error.

Also fix pre-existing clippy -D warnings failures (newer toolchain
lints, all in test code): int_plus_one (8 sites), bool_assert_comparison
(batching.rs), let_underscore on unit (manager.rs), redundant_locals
(g2_delete_race), identity_op (g1_review), unnecessary_cast
(g4_replay_freeze), disallowed eprintln allow (p3_4).

+++ tests/g4_replay_freeze.rs
@@ -28,7 +28,7 @@ fn legacy_freeze_replay_skips_already_spent_slot() {
 
     // Allocate a record region for one tx with one UTXO.
     let utxo_count = 1u32;
-    let record_size = TxMetadata::record_size_for(utxo_count) as u64;
+    let record_size = TxMetadata::record_size_for(utxo_count);
     let record_offset = alloc.allocate(record_size).unwrap();
 
     // Build a metadata header + one SPENT slot on the device.
@@ -100,7 +100,7 @@ fn legacy_freeze_replay_applies_on_unspent_slot() {
     let mut index = PrimaryBackend::new_in_memory(128).unwrap();
 
     let utxo_count = 1u32;
-    let record_size = TxMetadata::record_size_for(utxo_count) as u64;
+    let record_size = TxMetadata::record_size_for(utxo_count);
     let record_offset = alloc.allocate(record_size).unwrap();
 
     let mut meta = TxMetadata::new(utxo_count);
